# forest-back-demo/docs/integrations.openapi.yaml
openapi: 3.1.0
info:
  title: 산림재난 통합 연계 API
  version: 1.1.0
  description: |
    투비유니콘 통합 플랫폼과 장비·통신망·AI 제공업체 사이의 공통 계약입니다.
    외부 업체는 결과 수신 API를 호출하며, 투비유니콘은 등록된 endpointEnv를 통해
    외부 기능을 호출합니다. 이 명세는 외부 장비·RF·AI 자체 구현을 의미하지 않습니다.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /integrations/{capabilityId}/invoke:
    post:
      summary: 등록된 외부 장비 또는 AI API 호출
      description: 동일 capabilityId, direction, requestId 요청은 한 번만 처리합니다.
      operationId: invokeIntegration
      parameters:
        - $ref: "#/components/parameters/CapabilityId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/IntegrationEnvelope"}
      responses:
        "200": {description: 호출 완료 또는 기존 요청 상태 반환}
        "400": {description: 계약 검증 실패 또는 외부 호출 실패}
        "404": {description: 등록되지 않은 capabilityId}
  /integrations/{capabilityId}/results:
    post:
      summary: 외부 장비 텔레메트리 또는 AI 결과 수신
      operationId: receiveIntegrationResult
      parameters:
        - $ref: "#/components/parameters/CapabilityId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/IntegrationEnvelope"}
      responses:
        "200": {description: 저장 완료 또는 기존 요청 상태 반환}
        "400": {description: 계약 검증 또는 저장 실패}
        "404": {description: 등록되지 않은 capabilityId}
  /integrations/{capabilityId}/messages/{requestId}:
    get:
      summary: 요청 처리 상태와 오류 조회
      operationId: getIntegrationMessage
      parameters:
        - $ref: "#/components/parameters/CapabilityId"
        - name: requestId
          in: path
          required: true
          schema: {type: string, format: uuid}
        - name: direction
          in: query
          required: true
          schema: {type: string, enum: [INBOUND, OUTBOUND]}
      responses:
        "200":
          description: 접수·처리·실패 상태
        "404":
          description: 요청 없음
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CapabilityId:
      name: capabilityId
      in: path
      required: true
      schema: {type: string, minLength: 3, maxLength: 150}
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: context.requestId와 동일한 UUID
      schema: {type: string, format: uuid}
...
  schemas:
...
    IntegrationEnvelope:
      type: object
      required: [context, data]
      properties:
        context:
          type: object
          required: [eventId, requestId, sourceSystem, occurredAt, schemaVersion]
          properties:
            eventId: {type: string, format: uuid}
            requestId: {type: string, format: uuid}
            correlationId: {type: string, format: uuid}
            sourceSystem: {type: string, minLength: 2, maxLength: 100}
            reportedByAssetId:
              type: string
              format: uuid
              description: 통합 API를 실제 호출한 게이트웨이·GCS·NMS 자산 ID. reportingRole과 함께 사용한다.
            reportingRole:
              type: string
              enum: [GATEWAY, GCS, NMS, DEVICE, SERVICE]
              description: 실제 보고 주체의 역할. reportedByAssetId와 함께 사용한다.
            occurredAt: {type: string, format: date-time}
            sentAt: {type: string, format: date-time}
            schemaVersion: {type: string, const: "1.0"}
        data:
          type: object
          additionalProperties: true
...